+++ 84 Super Star Trek/csharp/Objects/Starbase.cs
@@ -6,21 +6,21 @@ namespace SuperStarTrek.Objects
     {
         private readonly Input _input;
         private readonly Output _output;
-        private readonly double _repairDelay;
+        private readonly float _repairDelay;
 
         public Starbase(Random random, Input input, Output output)
         {
-            _repairDelay = random.GetDouble() * 0.5;
+            _repairDelay = random.GetFloat() * 0.5f;
             _input = input;
             _output = output;
         }
 
         public override string ToString() => ">!<";
 
-        internal bool TryRepair(Enterprise enterprise, out double repairTime)
+        internal bool TryRepair(Enterprise enterprise, out float repairTime)
         {
-            repairTime = enterprise.DamagedSystemCount * 0.1 + _repairDelay;
-            if (repairTime >= 1) { repairTime = 0.9; }
+            repairTime = enterprise.DamagedSystemCount * 0.1f + _repairDelay;
+            if (repairTime >= 1) { repairTime = 0.9f; }
 
             _output.Write(Strings.RepairEstimate, repairTime);
             if (_input.GetYesNo(Strings.RepairPrompt, Input.YesNoMode.TrueOnY))
